Moved interfaces from interfaces/data/ to interfaces/
