+++ src/interfaces/wind-report.ts
