+++ src/interfaces/windfinder.ts
