+++ src/interfaces/windguru.ts
