+++ src/interfaces/windy.ts
